# apps/spire/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "vex-chat spire",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:{port}",
      "variables": {
        "port": {
          "default": "16777"
        }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/server": {
      "post": {
        "operationId": "createServer",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Server created"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/server/{id}": {
      "get": {
        "operationId": "getServer",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Server"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "operationId": "deleteServer",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/server/{id}/channels": {
      "post": {
        "operationId": "createChannel",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Channel created"
          }
        }
      },
      "get": {
        "operationId": "listChannels",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Channel list"
          }
        }
      }
    },
    "/channel/{id}": {
      "delete": {
        "operationId": "deleteChannel",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          }
        }
      }
    },
    "/server/{serverID}/permissions": {
      "get": {
        "operationId": "listPermissions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "serverID",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Permissions"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/permission/{id}": {
      "delete": {
        "operationId": "deletePermission",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          }
        }
      }
    },
    "/server/{serverID}/invites": {
      "post": {
        "operationId": "createInvite",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "serverID",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Invite created"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "operationId": "listInvites",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "serverID",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Invite list"
          }
        }
      }
    },
...
  },
...
}